service/requirements.txt
anyio==3.4.0
asgiref==3.4.1
attrs==21.3.0
autopep8==1.6.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
ecdsa==0.17.0
fastapi==0.70.1
h11==0.12.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pymongo==4.0.1
pyparsing==3.0.6
pytest==6.2.5
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
toml==0.10.2
typing-extensions==4.0.1
uvicorn==0.16.0
